Move the facade to dedicated namespaced class
