+++ packages/framework/src/Modules/DataCollections/DataCollectionServiceProvider.php
@@ -12,7 +12,7 @@ class DataCollectionServiceProvider extends ServiceProvider
     {
         // Register the class alias
         AliasLoader::getInstance()->alias(
-            'Collection', DataCollection::class
+            'Collection', Facades\Collection::class
         );
     }
